backend/requirements.txt
# ==========================================
# Investor-Report-Finder Backend Dependencies
# ==========================================

# Core Web Framework
fastapi>=0.115.2
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP & Web Scraping
requests>=2.32.3
beautifulsoup4>=4.12.0
httpx>=0.28.1

# LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0

# Search & Document Processing
tavily-python>=0.3.0
pdfplumber>=0.10.0
pandas>=2.0.0
numpy>=1.24.0

# Authentication & Security
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0

# Database - Supabase
supabase>=2.0.0

# Database - SQLAlchemy (for local cache)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Additional Dependencies
python-multipart>=0.0.18
tiktoken>=0.5.0
tenacity>=8.2.0
